refactor: introduce dedicated module for integrations
