+++ src/integration/mod.rs
@@ -1,0 +1,3 @@
+//! Integrations with external systems.
+pub mod taxer;
+pub mod universalbank;
