+++ src/lib.rs
@@ -11,8 +11,7 @@ pub mod domain;
 pub mod filter;
 pub mod income;
 pub mod init;
+pub mod integration;
 pub mod payment;
 pub mod report;
-pub mod taxer;
 pub mod time;
-pub mod universalbank;
